refactor(Coupon): 패키지명 변경
- 이전: payment
- 이후: coupon
