+++ app/src/main/java/woowacourse/shopping/data/coupon/remote/CouponApiService.kt
@@ -1,8 +1,8 @@
-package woowacourse.shopping.data.payment.remote
+package woowacourse.shopping.data.coupon.remote
 
 import retrofit2.Response
 import retrofit2.http.GET
-import woowacourse.shopping.data.payment.remote.dto.CouponResponse
+import woowacourse.shopping.data.coupon.remote.dto.CouponResponse
 
 interface CouponApiService {
     @GET("/coupons")
