+++ app/src/main/java/woowacourse/shopping/data/coupon/remote/datasource/CouponDataSource.kt
@@ -1,7 +1,7 @@
-package woowacourse.shopping.data.payment.remote.datasource
+package woowacourse.shopping.data.coupon.remote.datasource
 
 import woowacourse.shopping.data.common.ResponseResult
-import woowacourse.shopping.data.payment.remote.dto.CouponResponse
+import woowacourse.shopping.data.coupon.remote.dto.CouponResponse
 
 interface CouponDataSource {
     suspend fun loadCoupons(): ResponseResult<CouponResponse>
