+++ app/src/main/java/woowacourse/shopping/data/coupon/remote/datasource/CouponRemoteDataSource.kt
@@ -1,9 +1,9 @@
-package woowacourse.shopping.data.payment.remote.datasource
+package woowacourse.shopping.data.coupon.remote.datasource
 
 import woowacourse.shopping.data.common.ApiResponseHandler.handleApiResponse
 import woowacourse.shopping.data.common.ResponseResult
-import woowacourse.shopping.data.payment.remote.CouponApiService
-import woowacourse.shopping.data.payment.remote.dto.CouponResponse
+import woowacourse.shopping.data.coupon.remote.CouponApiService
+import woowacourse.shopping.data.coupon.remote.dto.CouponResponse
 
 class CouponRemoteDataSource(
     private val couponApiService: CouponApiService,
