+++ app/src/main/java/woowacourse/shopping/data/coupon/remote/dto/CouponResponse.kt
@@ -1,0 +1,5 @@
+package woowacourse.shopping.data.coupon.remote.dto
+
+data class CouponResponse(
+    val coupons: List<CouponDto>,
+)
