--- app/src/main/java/woowacourse/shopping/data/payment/remote/dto/CouponResponse.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.data.payment.remote.dto
-
-import woowacourse.shopping.data.payment.remote.dto.CouponDto
-
-data class CouponResponse(
-    val coupons: List<CouponDto>,
-)
